refactor(docs): tooltip
