+++ apps/docs/content/components/tooltip/colors.ts
@@ -1,30 +1,4 @@
-const App = `import {Tooltip, Button} from "@nextui-org/react";
-
-export default function App() {
-  const colors = [
-    "default",
-    "primary",
-    "secondary",
-    "success",
-    "warning",
-    "danger",
-    "foreground",
-  ];
-
-  return (
-    <div className="flex flex-wrap gap-4">
-      {colors.map((color) => {
-        return (
-          <Tooltip key={color} color={color} content={color} className="capitalize">
-            <Button variant="flat" color={color} className="capitalize">
-              {color}
-            </Button>
-          </Tooltip>
-        )
-      })}
-    </div>
-  );
-}`;
+import App from "./colors.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
