+++ apps/docs/content/components/tooltip/custom-styles.ts
@@ -1,26 +1,4 @@
-const App = `import {Tooltip, Button} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Tooltip 
-      showArrow
-      placement="right"
-      content="I am a tooltip"
-      classNames={{
-        base: [
-          // arrow color
-          "before:bg-neutral-400 dark:before:bg-white",
-        ],
-        content: [
-          "py-2 px-4 shadow-xl",
-          "text-black bg-gradient-to-br from-white to-neutral-400",
-        ],
-      }}
-    >
-      <Button variant="flat">Hover me</Button>
-    </Tooltip>
-  );
-}`;
+import App from "./custom-styles.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
