+++ apps/docs/content/components/tooltip/placements.ts
@@ -1,38 +1,4 @@
-const App = `import {Tooltip, Button} from "@nextui-org/react";
-
-export default function App() {
-  const placements = [
-    "top-start",
-    "top",
-    "top-end",
-    "bottom-start",
-    "bottom",
-    "bottom-end",
-    "left-start",
-    "left",
-    "left-end",
-    "right-start",
-    "right",
-    "right-end",
-  ];
-
-  return (
-    <div className="flex flex-wrap md:inline-grid md:grid-cols-3 gap-4">
-      {placements.map((placement) => (
-        <Tooltip
-          key={placement}
-          placement={placement}
-          content={placement}
-          color="secondary"
-        >
-          <Button variant="flat" color="secondary" className="capitalize">
-            {placement}
-          </Button>
-        </Tooltip>
-      ))}
-    </div>
-  );
-}`;
+import App from "./placements.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
