+++ apps/docs/content/components/tooltip/usage.raw.jsx
@@ -1,0 +1,9 @@
+import {Tooltip, Button} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <Tooltip content="I am a tooltip">
+      <Button>Hover me</Button>
+    </Tooltip>
+  );
+}
